Rename: 폴더 정리, /page.tsx 주석 제거
